menu printer has been updated, all exception catched

+++ dev/src/Business_Layer_Trucking/Delivery/Item.java
@@ -2,16 +2,16 @@ package Business_Layer_Trucking.Delivery;
 
 public class Item {
     private  int ID;
-    private int weight;
+    private double weight;
     private String name;
 
-    public Item(int id, int weight, String name){
+    public Item(int id, double weight, String name){
         this.ID=id;
         this.name=name;
         this.weight=weight;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
 
@@ -27,7 +27,7 @@ public class Item {
         return name;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
